0lith-obsidian-bridge/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-frontmatter>=1.1.0
requests>=2.31.0
qdrant-client>=1.7.0
